requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2.9
django-cors-headers==4.3.1
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
filelock==3.13.4
fsspec==2024.3.1
gunicorn==22.0.0
idna==3.6
intel-openmp==2021.4.0
Jinja2==3.1.3
joblib==1.4.0
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
sympy==1.12
tbb==2021.12.0
torch==2.3.0
torchvision==0.18.0
typing_extensions==4.11.0
tzdata==2023.4
urllib3==2.1.0
whitenoise==6.6.0
